+++ alerts/lib/deadman_alerttask.py
@@ -1,0 +1,9 @@
+from alerttask import AlertTask
+
+
+class DeadmanAlertTask(AlertTask):
+
+    def executeSearchEventsSimple(self):
+        # We override this method to specify the size as 1
+        # since we only care about if ANY events are found or not
+        return self.main_query.execute(self.es, indices=self.event_indices, size=1)