+++ rest/plugins/cymon.py
@@ -59,7 +59,7 @@ class message(object):
             response.status = 500
 
         print(requestDict, requestDict.keys())
-        if 'ipaddress' in requestDict.keys():
+        if 'ipaddress' in requestDict:
             url="https://cymon.io/api/nexus/v1/ip/{0}/events?combined=true&format=json".format(requestDict['ipaddress'])
 
             # add the cymon api key?