+++ tests/alerts/test_proxy_drop_non_standard_port.py
@@ -17,47 +17,33 @@ class TestAlertProxyDropNonStandardPort(AlertTestSuite):
     default_event = {
         "_type": "event",
         "_source": {
-            "category": "squid",
-            "tags": ["squid"],
+            "category": "proxy",
             "details": {
                 "sourceipaddress": "1.2.3.4",
-                "destination": "evil.com:6667",
-                "proxyaction": "TCP_DENIED/-",
-                "tcpaction": "CONNECT"
-            }
-        }
+                "destination": "evil.com:1337",
+                "destinationport": "1337",
+                "proxyaction": "TCP_DENIED",
+                "method": "CONNECT",
+            },
+        },
     }
 
     default_event2 = AlertTestSuite.copy(default_event)
-    default_event2["_source"]["details"]["destination"] = "evil.com:1337"
-
-    # This event is the default negative event that will cause the
-    # alert to trigger
-    default_negative_event = {
-        "_type": "event",
-        "_source": {
-            "category": "squid",
-            "tags": ["squid"],
-            "details": {
-                "sourceipaddress": "1.2.3.4",
-                "destination": "example.com:443",
-                "proxyaction": "TCP_DENIED/-",
-                "tcpaction": "CONNECT"
-            }
-        }
-    }
+    default_event2["_source"]["details"]["destination"] = "evil.com:6667"
+    default_event2["_source"]["details"]["destinationport"] = "6667"
 
     # This alert is the expected result from running this task
     default_alert = {
         "category": "squid",
-        "tags": ['squid', 'proxy'],
+        "tags": ["squid", "proxy"],
         "severity": "WARNING",
-        "summary": 'Suspicious Proxy DROP event(s) detected from 1.2.3.4 to the following non-std port destination(s): evil.com:6667'
+        "summary": "Suspicious Proxy DROP event(s) detected from 1.2.3.4 to the following non-std port destination(s): evil.com:1337",
     }
 
     default_alert_aggregated = AlertTestSuite.copy(default_alert)
     default_alert_aggregated[
-        "summary"] = 'Suspicious Proxy DROP event(s) detected from 1.2.3.4 to the following non-std port destination(s): evil.com:1337,evil.com:6667'
+        "summary"
+    ] = "Suspicious Proxy DROP event(s) detected from 1.2.3.4 to the following non-std port destination(s): evil.com:1337,evil.com:6667"
 
     test_cases = []
 
@@ -65,7 +51,7 @@ class TestAlertProxyDropNonStandardPort(AlertTestSuite):
         PositiveAlertTestCase(
             description="Positive test with default events and default alert expected",
             events=AlertTestSuite.create_events(default_event, 1),
-            expected_alert=default_alert
+            expected_alert=default_alert,
         )
     )
 
@@ -75,20 +61,23 @@ class TestAlertProxyDropNonStandardPort(AlertTestSuite):
         PositiveAlertTestCase(
             description="Positive test with default events and default alert expected",
             events=events1 + events2,
-            expected_alert=default_alert_aggregated
-        )
-    )
-
-    test_cases.append(
-        NegativeAlertTestCase(
-            description="Positive test with default events and default alert expected",
-            events=AlertTestSuite.create_events(default_negative_event, 1),
+            expected_alert=default_alert_aggregated,
         )
     )
 
     events = AlertTestSuite.create_events(default_event, 10)
     for event in events:
-        event['_source']['category'] = 'bad'
+        event["_source"]["details"]["destinationport"] = "443"
+    test_cases.append(
+        NegativeAlertTestCase(
+            description="Negative test case with events with whitelisted port",
+            events=events,
+        )
+    )
+
+    events = AlertTestSuite.create_events(default_event, 10)
+    for event in events:
+        event["_source"]["category"] = "bad"
     test_cases.append(
         NegativeAlertTestCase(
             description="Negative test case with events with incorrect category",
@@ -98,23 +87,14 @@ class TestAlertProxyDropNonStandardPort(AlertTestSuite):
 
     events = AlertTestSuite.create_events(default_event, 10)
     for event in events:
-        event['_source']['tags'] = 'bad tag example'
+        event["_source"][
+            "utctimestamp"
+        ] = AlertTestSuite.subtract_from_timestamp_lambda({"minutes": 241})
+        event["_source"][
+            "receivedtimestamp"
+        ] = AlertTestSuite.subtract_from_timestamp_lambda({"minutes": 241})
     test_cases.append(
         NegativeAlertTestCase(
-            description="Negative test case with events with incorrect tags",
-            events=events,
-        )
-    )
-
-    events = AlertTestSuite.create_events(default_event, 10)
-    for event in events:
-        event['_source']['utctimestamp'] = AlertTestSuite.subtract_from_timestamp_lambda({
-                                                                                         'minutes': 241})
-        event['_source']['receivedtimestamp'] = AlertTestSuite.subtract_from_timestamp_lambda({
-                                                                                              'minutes': 241})
-    test_cases.append(
-        NegativeAlertTestCase(
-            description="Negative test case with old timestamp",
-            events=events,
+            description="Negative test case with old timestamp", events=events
         )
     )