+++ tests/unit_test_suite.py
@@ -10,6 +10,7 @@ from dateutil.parser import parse
 
 import random
 import pytest
+import sys
 
 from mozdef_util.utilities import toUTC
 
@@ -44,6 +45,9 @@ class UnitTestSuite(object):
             self.reset_elasticsearch()
         if pytest.config.option.delete_queues:
             self.reset_rabbitmq()
+        # Remove any leftover plugin module as a result of loading
+        if 'plugins' in sys.modules:
+            del sys.modules['plugins']
 
     def populate_test_event(self, event, event_type='event'):
         self.es_client.save_event(body=event, doc_type=event_type)
@@ -67,8 +71,8 @@ class UnitTestSuite(object):
         self.es_client.delete_index(self.alert_index_name, True)
         self.es_client.delete_index('alerts', True)
 
-    def flush(self, index_name):
-        self.es_client.flush(index_name)
+    def refresh(self, index_name):
+        self.es_client.refresh(index_name)
 
     def random_ip(self):
         return str(random.randint(1, 255)) + "." + str(random.randint(1, 255)) + "." + str(random.randint(1, 255)) + "." + str(random.randint(1, 255))